typo

+++ sandbox/test_layout.cc
@@ -71,12 +71,12 @@ void compute_0(const std::vector<Matrix_t> & dIljn, const std::vector<Matrix_t> 
     auto & Yjm = Yljm[l];
     auto & R3nm = Rl3nm[l];
     const auto n_max = dIjn.cols();
-    const int m_max = 2*l+1;
+    const size_t m_max = 2*l+1;
     auto inter = Matrix_t(n_max, m_max);
     for (int i_neigh{0}; i_neigh < dIjn.rows(); i_neigh++) {
       for (int i_der{0}; i_neigh < 3; i_der++) {
         Eigen::array<int, 4> offsets = {i_neigh, i_der, 0, 0};
-        Eigen::array<int, 4> extents = {1, 1, n_max , m_max};
+        Eigen::array<int, 4> extents = {1, 1, n_max, m_max};
         inter = rij(i_neigh, i_der) * dIjn.row(i_neigh).transpose() * Yjm.row(i_neigh);
         auto inter_ten = Tnm_CMap_t(inter.data(), n_max, m_max);
         R3nm.slice(offsets, extents) = inter_ten;
@@ -123,8 +123,6 @@ int main(int argc, char * argv[]) {
     compute_0(dIljn, Yljm, rij, Rl3nm);
     elapsed[looper] = timer.elapsed();
   }
-
-
 
   json cmp0{};
   cmp0["elapsed_mean"] = elapsed.mean();
